+++ src/main/java/com/joojoo/api/user/application/port/in/SocialLoginUseCase.java
@@ -1,9 +1,9 @@
-package com.joojoo.api.user.application.auth;
+package com.joojoo.api.user.application.port.in;
 
 import com.joojoo.api.user.presentation.dto.request.AuthTokenDto;
 import com.joojoo.api.user.presentation.dto.response.LoginResponse;
 
-public interface AuthSocialService {
+public interface SocialLoginUseCase {
     LoginResponse kakaoWebSocialLogin(String code);
     LoginResponse kakaoAppSocialLogin(AuthTokenDto authTokenDto);
     LoginResponse appleSocialLogin(String code);
